+++ src/utils/ImageHandler.py
@@ -1,4 +1,3 @@
-
 import tkinter as tk
 from tkinter import filedialog
